+++ timePicker.go
@@ -118,7 +118,6 @@ func newTimePicker(session Session) View {
 func (picker *timePickerData) init(session Session) {
 	picker.viewData.init(session)
 	picker.tag = "TimePicker"
-	picker.hasHtmlDisabled = true
 	picker.normalize = normalizeTimePickerTag
 	picker.get = picker.getFunc
 	picker.set = picker.setFunc